--- bin/README.md
@@ -1,1 +1,0 @@
-# DailyConsume_BE
