--- bin/src/main/resources/application.properties
@@ -1,21 +1,0 @@
-spring.application.name=dailyconsume
-
-server.port=9090
-
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.generate-ddl=false
-spring.jpa.show-sql=true
-
-logging.level.com.shinhan.dailyconsume=info
-logging.level.org.springframework.security.web=trace
-
-spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
-spring.datasource.url=jdbc:mariadb://localhost:3306/dailyconsume
-spring.datasource.username=springuser
-spring.datasource.password=root
-spring.jpa.database-platform=org.hibernate.dialect.MariaDBDialect
-
-spring.servlet.multipart.enabled=true
-spring.servlet.multipart.location=C:\\upload
-spring.servlet.multipart.max-request-size=30MB
-spring.servlet.multipart.max-file-size=10MB
