--- src/main/java/com/shinhan/dailyconsume/domain/CardEntity.java
@@ -1,28 +1,0 @@
-package com.shinhan.dailyconsume.domain;
-
-import java.sql.Timestamp;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "t_card")
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Data
-public class CardEntity {
-	@Id
-	Long cardId;
-	String cardName;
-	String cardImgUrl;
-	String cardPageUrl;
-	
-}
