--- src/main/java/com/shinhan/dailyconsume/domain/TestEntity.java
@@ -1,25 +1,0 @@
-package com.shinhan.dailyconsume.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "t_test")
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Data
-public class TestEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-	
-	private String str;
-}
