+++ src/main/java/com/shinhan/dailyconsume/repository/TestRepository.java
@@ -2,7 +2,7 @@ package com.shinhan.dailyconsume.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.shinhan.dailyconsume.domain.TestEntity;
+import com.shinhan.dailyconsume.Domain.TestEntity;
 
 public interface TestRepository extends JpaRepository<TestEntity, Long>{
