+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.application.name=dailyconsume
 
-server.port=9999
+server.port=9090
 
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.generate-ddl=false
@@ -10,9 +10,9 @@ logging.level.com.shinhan.dailyconsume=info
 logging.level.org.springframework.security.web=trace
 
 spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
-spring.datasource.url=jdbc:mariadb://localhost:3307/firstzonedb
+spring.datasource.url=jdbc:mariadb://localhost:3306/dailyconsume
 spring.datasource.username=springuser
-spring.datasource.password=1234
+spring.datasource.password=root
 spring.jpa.database-platform=org.hibernate.dialect.MariaDBDialect
 
 spring.servlet.multipart.enabled=true
